OpenGL: Move Attributes enum to a more appropriate file
